[#21949] bad argument type for PETSc Mat
- did not compile on 32 bits platforms

+++ bibfor/elim_lagr/elg_comptt.F90
@@ -290,7 +290,7 @@ subroutine elg_comptt(c, t, nbeq, clag1)
 !
 !
 #else
-    integer(kind=8) :: c, t
+    integer :: c, t
     integer :: nbeq, clag1
     ASSERT(.false.)
 #endif
